Fix Rails path mutation
ActionDispatch will mutate the uri of a request before the application
actually gets a chance to see it. This can lead to auth failures where
the client signed with the original version of the path, and the server
is comparing against a mutated version.

Instead for ActionDispatch::Request we should use #original_fullpath,
which is passed to Rails from Rack.